# docs/swagger.yaml
definitions:
...
  class_request.ModifyClassMemberRequest:
    properties:
      class_id:
        type: integer
      student_id:
        items:
          type: integer
        type: array
    type: object
  class_request.ModifyClassRequest:
    properties:
      class_code:
        type: string
      class_name:
        type: string
    type: object
...
  class_response.DetailClassResponse:
    properties:
      class_code:
        $ref: '#/definitions/response.GeneralLabelKeyResponse'
      class_name:
        type: string
      id:
        type: integer
    type: object
...
  database.Paginator:
    properties:
      limit:
        type: integer
      nextPage:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      prevPage:
        type: integer
      records: {}
      totalPage:
        type: integer
      totalRecord:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  response.BaseResponse:
    properties:
      code:
        type: integer
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: string
    type: object
  response.GeneralLabelKeyResponse:
    properties:
      key: {}
      label:
        type: string
    type: object
  school.Class:
    properties:
      classCode:
        type: string
      className:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      detailClassCode:
        $ref: '#/definitions/school.ClassCode'
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  school.ClassCode:
    properties:
      class_member:
        items:
          $ref: '#/definitions/school.Class'
        type: array
      code:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  view.VStudent:
    properties:
      class_id:
        type: integer
      class_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      nisn:
        type: string
      user_id:
        type: integer
    type: object
host: api.sman16bekasi.id/gateway-service
info:
  contact:
    email: cbt@sman16bekasi.id
    name: CBT Team
    url: https://sman16bekasi.id
  description: This is documentation technical
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: CBT Documentation
  version: "1.0"
paths:
...
  /academic/class/{classId}/member:
    get:
      consumes:
      - application/json
      description: Member of Class
      parameters:
      - description: Class ID
        in: path
        name: classId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Member class response
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/view.VStudent'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: This endpoint about Member of class
      tags:
      - Class
  /academic/class/all:
    get:
      consumes:
      - application/json
      description: Return Class list data
      parameters:
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - default: 10
        description: 'Number of items per page (default: 10)'
        in: query
        name: size
        type: integer
      - description: Search key field (e.g., name, email, etc.)
        in: query
        name: search.key
        type: string
      - description: Search value for the specified key
        in: query
        name: search.value
        type: string
      - description: Filter conditions (varies depending on the API)
        in: query
        name: filter
        type: object
      produces:
      - application/json
      responses:
        "200":
          description: Pagination data class
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/database.Paginator'
                  - properties:
                      records:
                        $ref: '#/definitions/school.Class'
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: This endpoint about list of class
      tags:
      - Class
  /academic/class/create:
    post:
      consumes:
      - application/json
      description: Create New Class
      parameters:
      - description: Body Request of Create New Class
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class_request.ModifyClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Create class response
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/class_response.DetailClassResponse'
              type: object
      security:
      - BearerAuth: []
      summary: This endpoint about create new class
      tags:
      - Class
  /academic/class/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Class
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delete class response
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: This endpoint about Delete class
      tags:
      - Class
  /academic/class/detail/{id}:
    get:
      consumes:
      - application/json
      description: Return detail information class
      parameters:
      - description: Class ID
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detail of class
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/class_response.DetailClassResponse'
              type: object
      security:
      - BearerAuth: []
      summary: This endpoint about detail of class
      tags:
      - Class
  /academic/class/member/{id}/delete:
    delete:
      consumes:
      - application/json
      description: Member of Class
      parameters:
      - description: Request
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delete Member class response
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - BearerAuth: []
      summary: This endpoint about Member of class
      tags:
      - Class
  /academic/class/member/add:
    post:
      consumes:
      - application/json
      description: Add Member of Class
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class_request.ModifyClassMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Create Member class response
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/class_request.ModifyClassMemberRequest'
              type: object
      security:
      - BearerAuth: []
      summary: This endpoint about Add Member of class
      tags:
      - Class
...
  /academic/class/update/{id}:
    put:
      consumes:
      - application/json
      description: Update Class
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      - description: Body Request of Update Class
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class_request.ModifyClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update class response
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/class_response.DetailClassResponse'
              type: object
      security:
      - BearerAuth: []
      summary: This endpoint about Update class
      tags:
      - Class
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Use "Bearer {your_token}" to authenticate. Some endpoints are public
      and do not require authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...